Fix Level 7: Remove le_of_sq_le_sq from NewTheorem
Helper theorems defined in the same level file don't need to be listed in
NewTheorem - they're automatically available. Only theorems from OTHER files
need to be unlocked via NewTheorem.

This follows the pattern of other helper lemmas in Level07 (norm_nonzero_of_nonzero,
norm_sq_decomposition, etc.) which are not in NewTheorem but are available in the proof.